requirements.txt
absl-py==2.2.2
accelerate==1.7.0
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==25.3.0
autogluon==1.3.0
autogluon.common==1.3.0
autogluon.core==1.3.0
autogluon.features==1.3.0
autogluon.multimodal==1.3.0
autogluon.tabular==1.3.0
autogluon.timeseries==1.3.0
beartype==0.20.2
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
blis==1.3.0
boto3==1.38.20
botocore==1.38.20
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.8
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
coreforecast==0.0.15
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docx==0.2.4
einops==0.8.1
evaluate==0.4.3
executing==2.2.0
fastai==2.8.1
fastcore==1.8.2
fastdownload==0.0.7
fastprogress==1.0.3
fasttransform==0.0.2
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.0
fugue==0.9.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gluonts==0.16.1
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
hyperopt==0.2.7
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
liac-arff==2.5.0
lightgbm==4.6.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.15
mlforecast==0.13.6
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==3.4.2
nlpaug==1.1.11
nltk==3.9.1
numba==0.61.2
numpy==2.1.3
nvidia-ml-py3==7.352.0
ollama==0.4.9
omegaconf==2.3.0
openai==1.82.0
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
openml==0.15.1
openxlab==0.0.11
optuna==4.3.0
ordered-set==4.1.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.1
plum-dispatch==2.5.7
preshed==3.0.9
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
py4j==0.10.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
seqeval==1.2.2
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
statsforecast==2.0.1
statsmodels==0.14.4
streamlit==1.45.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenize_rt==6.2.0
tokenizers==0.21.1
toml==0.10.2
toolz==0.12.1
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triad==0.9.8
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
utilsforecast==0.2.10
virtualenv==20.31.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
window_ops==0.0.15
wrapt==1.17.2
xgboost==3.0.1
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.0
